agentic_sample_ad/requirements.txt
# 메인 에이전트 / 플래너
google-adk>=1.0.0

# A2A 원격 에이전트 호출 (event_manager)
a2a-sdk>=0.3.0
httpx>=0.27.0

# MCP 서버·클라이언트 (paper/sns 도구, Slack)
mcp>=1.0.0
pypdf>=5.0.0

# 테스트 시 참고: Google ADK/Gemini 사용을 위해 API 키 필요
# export GOOGLE_API_KEY=... 또는 GEMINI API 키 설정
